+++ src/GafferImage/Catalogue.cpp
@@ -615,9 +615,23 @@ void Catalogue::Image::copyFrom( const Image *other )
 
 Catalogue::Image::Ptr Catalogue::Image::load( const std::string &fileName )
 {
-	// Names of Plugs can not contain '.' - we want to load files like 'test.1001.exr', though
+	// GraphComponent names are much more restrictive than filenames, so
+	// we must replace all non-alphanumeric characters with `_`, and make
+	// sure it doesn't start with a number.
+	/// \todo Relax these restrictions and/or provide automatic name
+	/// sanitisation in GraphComponent.
 	std::string name = boost::filesystem::path( fileName ).stem().string();
-	boost::replace_all( name, ".", "_" );
+	std::replace_if(
+		name.begin(), name.end(),
+		[] ( char c ) {
+			return !std::isalnum( c, std::locale::classic() );
+		},
+		'_'
+	);
+	if( std::isdigit( name[0], std::locale::classic() ) )
+	{
+		name = "_" + name;
+	}
 
 	Ptr image = new Image( name, Plug::In, Plug::Default | Plug::Dynamic );
 	image->fileNamePlug()->setValue( fileName );